e2e tests: widen byte reads in fragment-style case literals
These cases are FRAGMENTS assembled with a prelude, so they carry no
`function main` and my extract-and-typecheck sweep skipped them. Each is a
byte read reaching i32 context: a bare `return e[2];` from a `: string`
local, a byte handed to `url_hex_val(c: i32)`, or one handed to
`step(m, c: i32)`.

Types were read from each declaration rather than pattern-matched -- the
closure array `hs: (() => i32)[]` and the `i32[]` / `i32[][]` `src`/`xs`
sites are untouched.

+++ internal/e2eselfhost/self_host_state_machine_ir_test.go
@@ -52,7 +52,7 @@ function main(): i32 {
     var s: string = "` + s + `";
     var m: M = M { state: Between, words: 0, chars: 0 };
     var i: i32 = 0;
-    while (i < s.len()) { m = step(m, s[i]); i = i + 1; }
+    while (i < s.len()) { m = step(m, s[i] as i32); i = i + 1; }
     return m.words * 10 + m.chars;
 }`
 }
